Move 'getlines' generator function from 'extract_reads.py' into bcftbx/utils.
