+++ bcftbx/utils.py
@@ -18,6 +18,10 @@ General utility classes:
   AttributeDictionary
   OrderedDictionary
 
+File reading utilities:
+
+  getlines
+
 File system wrappers and utilities:
 
   PathInfo
@@ -85,6 +89,13 @@ import grp
 import datetime
 import re
 import socket
+
+#######################################################################
+# Module constants
+#######################################################################
+
+# Default size of data to read from file
+CHUNKSIZE = 102400
 
 #######################################################################
 # General utility classes
@@ -196,6 +207,70 @@ class OrderedDictionary:
             self.__dict[key] = value
         else:
             raise KeyError, "Key '%s' already exists" % key
+
+#######################################################################
+# File reading utilities
+#######################################################################
+
+def getlines(filen):
+    """
+    Fetch lines from a file and return them one by one
+
+    This generator function tries to implement an efficient
+    method of reading lines sequentially from a file, by
+    minimising the number of reads from the file and
+    performing the line splitting in memory. It attempts
+    to replicate the idiom:
+
+    >>> for line in open(filen):
+    >>> ...
+
+    using:
+
+    >>> for line in getlines(filen):
+    >>> ...
+
+    The file can be gzipped; this function should handle
+    this invisibly provided that the file extension is
+    '.gz'.
+
+    Arguments:
+      filen (str): path of the file to read lines from
+
+    Yields:
+      String: next line of text from the file, with any
+        newline character removed.
+    """
+    if filen.split('.')[-1] == 'gz':
+        fp = gzip.open(filen,'rb')
+    else:
+        fp = open(filen,'rb')
+    # Read in data in chunks
+    buf = ''
+    lines = []
+    while True:
+        # Grab a chunk of data
+        data = fp.read(CHUNKSIZE)
+        # Check for EOF
+        if not data:
+            break
+        # Add to buffer and split into lines
+        buf = buf + data
+        if buf[0] == '\n':
+            buf = buf[1:]
+        if buf[-1] != '\n':
+            i = buf.rfind('\n')
+            if i == -1:
+                continue
+            else:
+                lines = buf[:i].split('\n')
+                buf = buf[i+1:]
+        else:
+            lines = buf[:-1].split('\n')
+            buf = ''
+        # Return the lines one at a time
+        for line in lines:
+            yield line
 
 #######################################################################
 # File system wrappers and utilities
